added support for JFR JDK API
